+++ src/components/social-media-urls/constants.ts
@@ -1,0 +1,28 @@
+import { URLProps } from "../../utils/models";
+
+export const socialMediaLinks: URLProps[] = [
+  {
+    url: "https://facebook.com/dewanshrawat15",
+    type: "facebook",
+  },
+  {
+    url: "https://github.com/dewanshrawat15",
+    type: "github",
+  },
+  {
+    url: "https://linkedin.com/in/dewanshrawat15",
+    type: "linkedin",
+  },
+  {
+    url: "https://instagram.com/dewanshrawat15",
+    type: "instagram",
+  },
+  {
+    url: "https://open.spotify.com/user/dewanshrawat15",
+    type: "spotify",
+  },
+  {
+    url: "mailto:dewanshrawat.work@gmail.com",
+    type: "mail",
+  },
+];
